+++ projects/_01_ingest/ascwds/fargate/validate_clean_ascwds_workplace.py
@@ -8,8 +8,50 @@ from polars_utils.validation.constants import GLOBAL_ACTIONS, GLOBAL_THRESHOLDS
 from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
     AscwdsWorkplaceCleanedColumns as ASCWPClean,
 )
+from utils.column_values.categorical_columns_by_dataset import (
+    ASCWDSWorkplaceCleanedCategoricalValues as CatValues,
+)
 
-COLS_TO_IMPORT = [ASCWPClean.establishment_id, ASCWPClean.ascwds_workplace_import_date]
+EXPECTED_SCHEMA = pb.Schema(
+    columns={
+        ASCWPClean.organisation_id: "String",
+        ASCWPClean.period: "String",
+        ASCWPClean.establishment_id: "String",
+        ASCWPClean.establishment_id_from_nmds: "String",
+        ASCWPClean.parent_id: "String",
+        ASCWPClean.nmds_id: "String",
+        ASCWPClean.establishment_created_date: "Date",
+        ASCWPClean.establishment_updated_date: "Date",
+        ASCWPClean.master_update_date: "Date",
+        ASCWPClean.last_logged_in_date: "Date",
+        ASCWPClean.la_permission: "String",
+        ASCWPClean.is_bulk_uploader: "String",
+        ASCWPClean.is_parent: "String",
+        ASCWPClean.parent_permission: "String",
+        ASCWPClean.registration_type: "String",
+        ASCWPClean.provider_id: "String",
+        ASCWPClean.location_id: "String",
+        ASCWPClean.establishment_type: "String",
+        ASCWPClean.establishment_name: "String",
+        ASCWPClean.address: "String",
+        ASCWPClean.postcode: "String",
+        ASCWPClean.region_id: "String",
+        ASCWPClean.total_staff: "Int32",
+        ASCWPClean.worker_records: "Int32",
+        ASCWPClean.total_starters: "String",
+        ASCWPClean.total_leavers: "String",
+        ASCWPClean.total_vacancies: "String",
+        ASCWPClean.main_service_id: "String",
+        ASCWPClean.version: "String",
+        ASCWPClean.ascwds_workplace_import_date: "Date",
+        ASCWPClean.master_update_date_org: "Date",
+        ASCWPClean.purge_date: "Date",
+        ASCWPClean.data_last_amended_date: "Date",
+        ASCWPClean.workplace_last_active_date: "Date",
+        ASCWPClean.total_staff_bounded: "Int32",
+        ASCWPClean.worker_records_bounded: "Int32",
+    }
+)
 
 
 def main(bucket_name: str, source_path: str, reports_path: str) -> None:
@@ -25,7 +67,6 @@ def main(bucket_name: str, source_path: str, reports_path: str) -> None:
     """
     source_df = utils.read_parquet(
         source=f"s3://{bucket_name}/{source_path}",
-        selected_columns=COLS_TO_IMPORT,
     )
 
     validation = (
@@ -36,10 +77,93 @@ def main(bucket_name: str, source_path: str, reports_path: str) -> None:
             brief=True,
             actions=GLOBAL_ACTIONS,
         )
+        # dataset schema
+        .col_schema_match(
+            schema=EXPECTED_SCHEMA,
+            brief="Dataset should match the expected schema",
+        )
         # index columns
         .rows_distinct(
             [ASCWPClean.establishment_id, ASCWPClean.ascwds_workplace_import_date]
-        ).interrogate()
+        )
+        # complete columns
+        .col_vals_not_null(
+            columns=[
+                ASCWPClean.organisation_id,
+                ASCWPClean.ascwds_workplace_import_date,
+                ASCWPClean.establishment_id,
+            ],
+            brief="Key columns should contain no null values",
+        )
+        # numerical
+        .col_vals_between(
+            columns=[
+                ASCWPClean.total_staff_bounded,
+                ASCWPClean.worker_records_bounded,
+            ],
+            left=1,
+            right=3000,
+            na_pass=True,
+            brief="Counts should be between 1 and 3000 where present.",
+        )
+        # categorical
+        # .col_vals_in_set(
+        #     ASCWPClean.establishment_type,
+        #     [*CatValues.establishment_type_column_values.categorical_values, None],
+        # )
+        # .col_vals_in_set(
+        #     ASCWPClean.parent_permission,
+        #     [*CatValues.parent_permission_column_values.categorical_values, None],
+        # )
+        .col_vals_in_set(
+            ASCWPClean.is_parent,
+            [*CatValues.is_parent_column_values.categorical_values, None],
+        )
+        # .col_vals_in_set(
+        #     ASCWPClean.main_service_id,
+        #     [*CatValues.main_service_id_column_values.categorical_values, None],
+        # )
+        # .col_vals_in_set(
+        #     ASCWPClean.registration_type,
+        #     [*CatValues.registration_type_column_values.categorical_values, None],
+        # )
+        # distinct values
+        # .specially(
+        #     vl.is_unique_count_equal(
+        #         ASCWPClean.establishment_type,
+        #         CatValues.establishment_type_column_values.count_of_categorical_values,
+        #     ),
+        #     brief=f"{ASCWPClean.establishment_type} should have exactly {CatValues.establishment_type_column_values.count_of_categorical_values} distinct values",
+        # )
+        # .specially(
+        #     vl.is_unique_count_equal(
+        #         ASCWPClean.parent_permission,
+        #         CatValues.parent_permission_column_values.count_of_categorical_values,
+        #     ),
+        #     brief=f"{ASCWPClean.parent_permission} should have exactly {CatValues.parent_permission_column_values.count_of_categorical_values} distinct values",
+        # )
+        .specially(
+            vl.is_unique_count_equal(
+                ASCWPClean.is_parent,
+                CatValues.is_parent_column_values.count_of_categorical_values,
+            ),
+            brief=f"{ASCWPClean.is_parent} should have exactly {CatValues.is_parent_column_values.count_of_categorical_values} distinct values",
+        )
+        # .specially(
+        #     vl.is_unique_count_equal(
+        #         ASCWPClean.main_service_id,
+        #         CatValues.main_service_id_column_values.count_of_categorical_values,
+        #     ),
+        #     brief=f"{ASCWPClean.main_service_id} should have exactly {CatValues.main_service_id_column_values.count_of_categorical_values} distinct values",
+        # )
+        # .specially(
+        #     vl.is_unique_count_equal(
+        #         ASCWPClean.registration_type,
+        #         CatValues.registration_type_column_values.count_of_categorical_values,
+        #     ),
+        #     brief=f"{ASCWPClean.registration_type} should have exactly {CatValues.registration_type_column_values.count_of_categorical_values} distinct values",
+        # )
+        .interrogate()
     )
     vl.write_reports(validation, bucket_name, reports_path)