+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/fargate/utils/merge_utils.py
@@ -1,1 +1,25 @@
 from polars_utils import utils
+
+
+def create_list_of_cols_for_ascwds():
+    """
+    Placeholder function to create a list of columns for the ASCWDS dataset."""
+    pass
+
+
+def convert_ascwds_job_role_columns_to_rows():
+    """
+    Placeholder function to convert ASCWDS job role columns to rows."""
+    pass
+
+
+def join_datasets():
+    """
+    Placeholder function to join the datasets."""
+    pass
+
+
+def apply_employment_status_magic_numbers():
+    """
+    Placeholder function to apply employment status magic numbers."""
+    pass