+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/tests/fargate/utils/test_merge_utils.py
@@ -1,6 +1,23 @@
-import unittest
-from unittest.mock import ANY, Mock, patch
-
 import projects._07_workforce_characteristics._01_starters_leavers_vacancies.fargate.utils.merge_utils as job
 
 PATCH_PATH = "projects._07_workforce_characteristics._01_starters_leavers_vacancies.fargate.utils.merge_utils"
+
+
+class TestCreateListOfColsForAscwds:
+    def test_create_list_of_cols_for_ascwds(self):
+        pass
+
+
+class TestConvertAscwdsJobRoleColumnsToRows:
+    def test_convert_ascwds_job_role_columns_to_rows(self):
+        pass
+
+
+class TestJoinDatasets:
+    def test_join_datasets(self):
+        pass
+
+
+class TestApplyEmploymentStatusMagicNumbers:
+    def test_apply_employment_status_magic_numbers(self):
+        pass